move homemap back to the page route
